Separate Rewarded-Video logic from Interstitial class.
